+++ app/build.gradle
@@ -7,12 +7,12 @@ plugins {
 
 android {
     namespace 'com.poklad.androidifystore'
-    compileSdk 33
+    compileSdk 34
 
     defaultConfig {
         applicationId "com.poklad.androidifystore"
         minSdk 24
-        targetSdk 33
+        targetSdk 34
         versionCode 1
         versionName "1.0"
 
@@ -46,12 +46,12 @@ dependencies {
     kapt 'com.github.bumptech.glide:compiler:4.12.0'
 
     //di
-    implementation 'com.google.dagger:dagger-android:2.35.1'
-    kapt 'androidx.annotation:annotation:1.6.0'
+    //dagger
+    implementation 'com.google.dagger:dagger:2.47'
     kapt 'com.google.dagger:dagger-compiler:2.47'
 
     //rx
-    implementation 'io.reactivex.rxjava2:rxjava:2.2.15'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.19'
     implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
     implementation 'com.squareup.retrofit2:adapter-rxjava2:2.9.0'
