+++ app/src/main/java/com/poklad/androidifystore/StoreApp.kt
@@ -2,6 +2,7 @@ package com.poklad.androidifystore
 
 import android.app.Application
 import com.poklad.androidifystore.di.AppComponent
+import com.poklad.androidifystore.di.DaggerAppComponent
 
 class StoreApp : Application() {
 
@@ -10,7 +11,7 @@ class StoreApp : Application() {
     }
 
     private fun initializationComponent(): AppComponent {
-        TODO()
+        return DaggerAppComponent.factory().create(this)
     }
 
 }
