+++ app/src/main/java/com/poklad/androidifystore/data/remote/StoreApi.kt
@@ -1,26 +1,26 @@
 package com.poklad.androidifystore.data.remote
 
-import com.poklad.androidifystore.data.remote.dto.ProductCategoryDto
-import com.poklad.androidifystore.data.remote.dto.ProductDtoItem
+import com.poklad.androidifystore.data.remote.model.ProductCategoryModel
+import com.poklad.androidifystore.data.remote.model.ProductItemModel
 import retrofit2.http.GET
 import retrofit2.http.Path
 
 interface StoreApi {
     @GET("products")
-    suspend fun getAllProducts(): List<ProductDtoItem>
+    suspend fun getAllProducts(): List<ProductItemModel>
 
     @GET("products/{productId}")
     suspend fun getProductById(
         @Path("productId")
         productId: Long
-    ): ProductDtoItem
+    ): ProductItemModel
 
     @GET("products/category/{categoryName}")
     suspend fun getProductsByCategories(
         @Path("categoryName")
         categoryName: String
-    ): List<ProductDtoItem>
+    ): List<ProductItemModel>
 
     @GET("products/categories")
-    suspend fun getAllCategories(): List<ProductCategoryDto>
+    suspend fun getAllCategories(): List<ProductCategoryModel>
 }
