--- app/src/main/java/com/poklad/androidifystore/data/remote/dto/ProductCategoryDto.kt
@@ -1,5 +1,0 @@
-package com.poklad.androidifystore.data.remote.dto
-
-data class ProductCategoryDto(
-    val category: String
-)
