+++ app/src/main/java/com/poklad/androidifystore/data/remote/model/ProductCategoryModel.kt
@@ -1,0 +1,11 @@
+package com.poklad.androidifystore.data.remote.model
+
+data class ProductCategoryModel(
+    val category: ProductCategory
+)
+enum class ProductCategory {
+    ELECTRONICS,
+    JEWELRY,
+    MENS_CLOTHING,
+    WOMENS_CLOTHING
+}
