--- app/src/main/java/com/poklad/androidifystore/domain/model/ProductCategory.kt
@@ -1,8 +1,0 @@
-package com.poklad.androidifystore.domain.model
-
-enum class ProductCategory {
-        ELECTRONICS,
-        JEWELRY,
-        MENS_CLOTHING,
-        WOMENS_CLOTHING
-}
