+++ app/src/main/java/com/poklad/androidifystore/domain/repositories/ProductsRepository.kt
@@ -2,7 +2,6 @@ package com.poklad.androidifystore.domain.repositories
 
 import com.poklad.androidifystore.domain.model.ProductItem
 
-//todo flow тут може бути, а вот livedata єто чисто во вьюмодел? А ресурс уже в UseCase?
 interface ProductsRepository {
     suspend fun getAllProducts(): List<ProductItem>
     suspend fun getProductsById(productId: Long): ProductItem
